refac (WinningLottoNUmbers, LottoNumbers) : List<LottoNumber> 를 LottoNumbers 일급 콜렉션 객체로 변환 리팩토링
