+++ src/main/java/lotto/model/LottoNumbers.java
@@ -1,12 +1,15 @@
 package lotto.model;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
+import lotto.code.ErrorCode;
+import lotto.exception.LottoException;
 import lotto.util.RandomUtil;
 import lotto.util.SplitUtil;
 
@@ -28,6 +31,8 @@ public class LottoNumbers {
 	}
 
 	private List<LottoNumber> generateLottoNumberList(String inputNumber) {
+		isNotLottoNumberSize(inputNumber);
+		isDuplicateLottoNumber(inputNumber);
 		List<LottoNumber> lottoNumberList = new ArrayList<>();
 
 		for (String number : SplitUtil.splitInputNumbers(inputNumber)) {
@@ -54,6 +59,20 @@ public class LottoNumbers {
 
 	public int size() {
 		return lottoNumberList.size();
+	}
+
+	private void isNotLottoNumberSize(String input) {
+		if (SplitUtil.splitInputNumbers(input).length != LottoNumbers.LOTTO_NUMBERS_SIZE) {
+			throw new LottoException(ErrorCode.IS_NOT_LOTTO_NUMBER_SIZE_ERROR);
+		}
+	}
+
+	private void isDuplicateLottoNumber(String input) {
+		Set<String> lottoNumberSet = new HashSet<>(Arrays.asList(SplitUtil.splitInputNumbers(input)));
+
+		if (lottoNumberSet.size() != LottoNumbers.LOTTO_NUMBERS_SIZE) {
+			throw new LottoException(ErrorCode.LOTTO_NUMBER_DUPLICATE_ERROR);
+		}
 	}
 
 	@Override
